[refactor] 마이페이지 패키지 변경, 비밀번호 변경 구분
