+++ src/main/java/com/triptune/domain/member/dto/request/ResetPasswordRequest.java
@@ -1,0 +1,36 @@
+package com.triptune.domain.member.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ResetPasswordRequest {
+
+    @NotBlank(message = "비밀번호 변경 토큰은 필수 입력 값입니다.")
+    private String passwordToken;
+
+    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
+    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,15}$", message = "비밀번호는 8자 이상 15자 이하의 영문, 숫자, 특수문자 조합이어야 합니다.")
+    private String password;
+
+    @NotBlank(message = "비밀번호 재입력은 필수 입력 값입니다.")
+    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,15}$", message = "비밀번호는 8자 이상 15자 이하의 영문, 숫자, 특수문자 조합이어야 합니다.")
+    private String rePassword;
+
+    @Builder
+    public ResetPasswordRequest(String passwordToken, String password, String rePassword) {
+        this.passwordToken = passwordToken;
+        this.password = password;
+        this.rePassword = rePassword;
+    }
+
+    public boolean isMatchPassword() {
+        return this.password.equals(this.rePassword);
+    }
+}
