+++ src/main/java/com/triptune/global/config/SecurityConstants.java
@@ -4,7 +4,7 @@ public class SecurityConstants {
 
     public static final String[] AUTH_WHITELIST = {
             "/swagger-ui/**", "/v3/api-docs/**", "/api/members/join", "/api/members/login", "/api/members/logout",
-            "/api/members/refresh", "/api/members/find-id", "/api/members/find-password", "/api/members/change-password",
+            "/api/members/refresh", "/api/members/find-id", "/api/members/find-password", "/api/members/reset-password",
             "/api/emails/**", "/api/travels/**", "/h2-console/**", "/", "/error", "/ws"
     };
