--- src/test/java/com/triptune/domain/mypage/MyPageTest.java
@@ -1,14 +1,0 @@
-package com.triptune.domain.mypage;
-
-import com.triptune.domain.BaseTest;
-import com.triptune.domain.mypage.dto.request.MyPagePasswordRequest;
-
-public class MyPageTest extends BaseTest {
-
-    protected MyPagePasswordRequest createMyPagePasswordRequest(String nowPassword, String newPassword, String rePassword){
-        return MyPagePasswordRequest.builder()
-                .nowPassword(nowPassword)
-                .newPassword(newPassword)
-                .rePassword(rePassword).build();
-    }
-}
